chore: update oazapfts and refactor code generation
- Updated oazapfts dependency to v7.5.0-alpha.3 in package.json and yarn.lock.
- Introduced new utility functions for operation name generation and deep object support.
- Removed the deprecated oazapfts-compat.ts file and refactored generate.ts to utilize the new public API.
- Enhanced type generation logic to support TypeScript enums based on the useEnumType option.
- Updated tests to reflect changes in type generation and added new tests for enum generation.
