+++ packages/rtk-query-codegen-openapi/package.json
@@ -79,9 +79,10 @@
   },
   "dependencies": {
     "@apidevtools/swagger-parser": "^10.1.1",
+    "@oazapfts/resolve": "^1.0.0",
     "commander": "^6.2.0",
     "lodash.camelcase": "^4.3.0",
-    "oazapfts": "^7.0.0",
+    "oazapfts": "7.5.0-alpha.3",
     "prettier": "^3.2.5",
     "semver": "^7.3.5",
     "swagger2openapi": "^7.0.4",
