+++ packages/rtk-query-codegen-openapi/src/generators/react-hooks.ts
@@ -1,5 +1,5 @@
 import ts from 'typescript';
-import { getOperationName } from '../oazapfts-compat';
+import { getOperationName } from '../utils';
 import { capitalize, isQuery } from '../utils';
 import type { OperationDefinition, EndpointOverrides, ConfigFile } from '../types';
 import { getOverrides } from '../generate';
